# Efficient package calculation