--- a/test/.gitignore
+++ b/test/.gitignore
@@ -2,3 +2,4 @@
 /*/work
 /*/.bob-*
 /*/log.txt
+/*/log-cmd.txt