--- a/test/multiPackage-clash/run.sh
+++ b/test/multiPackage-clash/run.sh
@@ -2,7 +2,4 @@
 . ../test-lib.sh 2>/dev/null || { echo "Must run in script directory!" ; exit 1 ; }
 
 # bob must fail because of package name clash
-if run_bob ls > log-cmd.txt 2>&1 ; then
-	echo "Parsing must fail!"
-	exit 1
-fi
+expect_fail run_bob ls