Restructure the internal package data structures and speed up initial calculation.

@kolewu @ThomasFeher Could you please provide some feedback if this works for you or breaks something on your side?

At the same time this cleans up the direct/indirect dependencies semantic. It is now possible to build packages that are only there due to "provideDeps" of downstream packages.

Fixes #98