@kolewu 2017-01-26
I have made some tests with this branch and found no problems. The initial recipe parsing for `bob ls` is longer than in the old version, but all following calls to `bob ls` are faster.

For our recipes (82 root recipes that recursively expand to 3223 with `bob ls -ra | wc -l`) the numbers for `bob ls` are:

Without cache:
```
Old version:          New version:
--------------------------------------
real	0m1,838s      real    0m3,851s
user	0m1,708s      user    0m3,700s
sys	0m0,036s      sys     0m0,060s
```

With cache:
```
Old version:          New version:
--------------------------------------
real	0m1,122s      real    0m0,237s
user	0m1,092s      user    0m0,216s
sys	0m0,024s      sys     0m0,012s    
```

So from our side and from the outside (I haven't reviewed the changes) it looks fine.