@jkloetzke 2017-01-27
Thanks for testing. This is looking promising. :smiley: 

Just as a side note. We have just 28 root recipes but they expand to staggering 663853 packages. Without the cache the new version is already miles ahead:

```
Old version:          New version:
--------------------------------------
real    1m44.313s     real    0m15.800s
user    1m43.796s     user    0m15.612s
sys     0m0.420s      sys     0m0.140s
```

with cache the difference is also huge:

```
Old version:          New version:
--------------------------------------
real    0m7.653s      real    0m0.886s
user    0m7.364s      user    0m0.832s
sys     0m0.288s      sys     0m0.048s
```